Reorganised new search structure
