+++ components/search/search.module.ts
@@ -3,7 +3,8 @@ import {CommonModule} from "@angular/common";
 import {FormsModule} from "@angular/forms";
 import {SuiDropdownModule} from "../dropdown/dropdown.module";
 import {SuiTransitionModule} from "../transition/transition.module";
-import {SuiSearch, SuiSearchValueAccessor} from './new-search';
+import {SuiSearch, SuiSearchValueAccessor} from './search';
+import {SearchService} from './search.service';
 
 @NgModule({
     imports: [
@@ -21,3 +22,5 @@ import {SuiSearch, SuiSearchValueAccessor} from './new-search';
     ]
 })
 export class SuiSearchModule {}
+
+export {SearchService};
