+++ components/search/search.ts
@@ -1,191 +1,146 @@
-import {Component, Directive, HostListener, HostBinding, ElementRef, AfterViewInit, ViewChild, EventEmitter, forwardRef} from '@angular/core';
+import {Component, ViewChild, HostBinding, Input, AfterViewInit, HostListener, EventEmitter, Output, forwardRef, Directive} from '@angular/core';
+import {DropdownService} from '../dropdown/dropdown.service';
+import {SuiDropdownMenu} from '../dropdown/dropdown-menu';
 import {NG_VALUE_ACCESSOR, ControlValueAccessor} from '@angular/forms';
-
-import {SuiDropdownMenu} from "../dropdown/dropdown-menu";
-import {Input, Output} from "@angular/core";
-import {SuiSearchService} from "./search.service";
+import {SearchService, LookupFn} from './search.service';
+import {RecursiveObject, readValue} from '../util/util';
 
 @Component({
     selector: 'sui-search',
     template: `
-<div class="ui icon input">
+<div class="ui input" [class.icon]="hasIcon">
     <input class="prompt" type="text" [attr.placeholder]="placeholder" autocomplete="off" [(ngModel)]="query">
-    <i *ngIf="icon" class="search icon"></i>
+    <i *ngIf="hasIcon" class="search icon"></i>
   </div>
-<div class="results" suiDropdownMenu>
-    <a class="result" *ngFor="let r of results; let i = index" (click)="select(r)">
-        <div *ngIf="highlightMatch" class="title" [innerHTML]="highlight(result(i))"></div>
-        <div *ngIf="!highlightMatch" class="title">{{ result(i) }}</div>
+<div class="results" suiDropdownMenu transition="scale" selectedItemClass="active">
+    <a class="result item" *ngFor="let r of results" (click)="select(r)">
+        <span *ngIf="!searchService.optionsLookup" [innerHTML]="searchService.highlightMatches(r)"></span>
+        <span *ngIf="searchService.optionsLookup">{{ readValue(r) }}</span>
     </a>
-    <div *ngIf="!results.length" class="message empty">
+    <div *ngIf="results.length == 0" class="message empty">
         <div class="header">No Results</div>
         <div class="description">Your search returned no results.</div>
     </div>
 </div>
-`
+`,
+    styles: [`
+:host {
+    display: inline-block;
+}
+`]
 })
-export class SuiSearch implements AfterViewInit {
+export class SuiSearch<T extends RecursiveObject> implements AfterViewInit {
+    public dropdownService:DropdownService;
+    public searchService:SearchService<T>;
+
     @ViewChild(SuiDropdownMenu)
-    private _dropdownMenu:SuiDropdownMenu;
-    // private _dropdownService:SuiDropdownService = new SuiDropdownService();
-    private _searchService:SuiSearchService = new SuiSearchService();
-
-    public selectedOption:any;
-
-    constructor(el:ElementRef) {
-        // this._dropdownService.dropdownElement = el;
-        // this._dropdownService.itemClass = "result";
-        // this._dropdownService.itemSelectedClass = "active";
-
-        // this._dropdownService.isOpenChange
-        //     .subscribe( (isOpen: boolean) => {
-        //         if (isOpen) {
-        //             if (!this._dropdownService.selectedItem) {
-        //                 this._dropdownService.selectNextItem();
-        //             }
-        //         }
-        //     });
-
-        // this._searchService.onSearchCompleted
-        //     .subscribe(() => {
-        //         this._dropdownService.isOpen = true;
-        //     });
-    }
+    private _menu:SuiDropdownMenu;
 
     @HostBinding('class.ui')
     @HostBinding('class.search')
-    public searchClasses = true;
+    private _searchClasses:boolean;
 
-    @Input()
-    public placeholder:string = "Search...";
-
-    @Input()
-    public get searchDelay() {
-        return this._searchService.searchDelay;
-    }
-
-    public set searchDelay(value:number) {
-        this._searchService.searchDelay = value;
+    @HostBinding('class.active')
+    public get isActive() {
+        return this.dropdownService.isOpen;
     }
 
     @Input()
-    public icon:boolean = true;
+    public hasIcon:boolean;
 
     @Input()
-    public highlightMatch:boolean = true;
+    public placeholder:string;
+
+    public get query() {
+        return this.searchService.query;
+    }
+
+    public set query(query:string) {
+        this.selectedItem = null;
+        this.searchService.updateQueryDelayed(query, () => this.dropdownService.setOpenState(this.searchService.query.length > 0));
+    }
 
     @Input()
-    public get optionsField() {
-        return this._searchService.optionsField;
+    public set options(options:T[] | LookupFn<T>) {
+        if (typeof(options) == "function") {
+            this.searchService.optionsLookup = options;
+            return;
+        }
+        this.searchService.options = options;
     }
 
-    public set optionsField(value:string) {
-        this._searchService.optionsField = value;
+    @Input()
+    public set optionsField(field:string) {
+        this.searchService.optionsField = field;
     }
 
-    @Output()
-    public selectedOptionChange:EventEmitter<any> = new EventEmitter<any>();
-
-    @Output()
-    public onItemSelected:EventEmitter<any> = new EventEmitter<any>();
+    @Input()
+    public set searchDelay(delay:number) {
+        this.searchService.searchDelay = delay;
+    }
 
     @HostBinding('class.loading')
-    public get loading() {
-        return this._searchService.loading;
+    public get isSearching() {
+        return this.searchService.isSearching;
     }
 
-    // @HostBinding('class.visible')
-    // public get isVisible() {
-    //     return this._dropdownService.transition.isVisible;
-    // }
-
-    // @HostBinding('class.active')
-    // @Input()
-    // public get isOpen():boolean {
-    //     return this._dropdownService.isOpen;
-    // }
-
-    // public set isOpen(value:boolean) {
-    //     this._dropdownService.isOpen = value;
-    // }
-
-    // @HostBinding('class.disabled')
-    // @Input()
-    // public get isDisabled():boolean {
-    //     return this._dropdownService.isDisabled;
-    // }
-
-    // public set isDisabled(value:boolean) {
-    //     this._dropdownService.isDisabled = value;
-    // }
-
-    @Input()
-    public get options():any {
-        return this._searchService.options;
+    public get results() {
+        return this.searchService.results;
     }
 
-    public set options(value:any) {
-        this._searchService.options = value;
+    public selectedItem:T;
+
+    @Output()
+    public onItemSelected:EventEmitter<T>;
+
+    public get ngModelChange() {
+        return this.onItemSelected;
     }
 
-    private get query():string {
-        return this._searchService.query;
+    constructor() {
+        this.dropdownService = new DropdownService();
+        this.searchService = new SearchService<T>();
+
+        this._searchClasses = true;
+        this.hasIcon = true;
+        this.placeholder = "Search...";
+
+        this.searchDelay = 200;
+
+        this.onItemSelected = new EventEmitter<T>();
     }
 
-    private set query(value:string) {
-        this._searchService.updateQuery(value);
+    public ngAfterViewInit() {
+        this._menu.service = this.dropdownService;
     }
 
-    private get results():Array<any> {
-        return this._searchService.results;
+    public select(item:T) {
+        this.writeValue(item);
+        this.onItemSelected.emit(item);
     }
 
-    private search():void {
-        this._searchService.search();
+    @HostListener("click", ['$event'])
+    public onClick(e:MouseEvent) {
+        e.stopPropagation();
+
+        if (this.searchService.query.length > 0) {
+            this.dropdownService.setOpenState(true);
+        }
     }
 
-    private result(i:number):any {
-        return this._searchService.readValue(this.results[i]);
+    public readValue(object:T) {
+        return readValue(object, this.searchService.optionsField);
     }
 
-    public highlight(result:string){
-        return result.replace(new RegExp(this.query,'i'), function(str) {return '<mark>'+str+'</mark>'});
-    }
-
-    public select(result:any):void {
-        this.selectedOption = result;
-        this.selectedOptionChange.emit(result);
-        this.onItemSelected.emit(result);
-        this._searchService.updateQuery(this._searchService.readValue(result), false);
-        // this._dropdownService.isOpen = false;
-    }
-
-    public writeValue(value:any) {
-        this.selectedOption = value;
-        this._searchService.updateQuery(this._searchService.readValue(value), false);
-    }
-
-    public ngAfterViewInit():void {
-        // this._dropdownMenu.service = this._dropdownService;
-    }
-
-    @HostListener('click', ['$event'])
-    public click(event:MouseEvent):boolean {
-        event.stopPropagation();
-
-        // if (!this._dropdownService.menuElement.nativeElement.contains(event.target)){
-        //     if (!this.isOpen && this.query) {
-        //         if (this.results.length) {
-        //             this.isOpen = true;
-        //         }
-        //         this.search();
-        //     }
-        // }
-        return false;
+    public writeValue(item:T) {
+        if (item) {
+            this.selectedItem = item;
+            this.searchService.updateQuery(this.readValue(item) as string, () => {});
+        }
     }
 }
 
-export const CUSTOM_VALUE_ACCESSOR: any = {
+export const SELECT_VALUE_ACCESSOR:any = {
     provide: NG_VALUE_ACCESSOR,
     useExisting: forwardRef(() => SuiSearchValueAccessor),
     multi: true
@@ -193,21 +148,25 @@ export const CUSTOM_VALUE_ACCESSOR: any = {
 
 @Directive({
     selector: 'sui-search',
-    host: {'(selectedOptionChange)': 'onChange($event)'},
-    providers: [CUSTOM_VALUE_ACCESSOR]
+    host: {
+        '(onItemSelected)': 'onChange($event)'
+    },
+    providers: [SELECT_VALUE_ACCESSOR]
 })
-export class SuiSearchValueAccessor implements ControlValueAccessor {
+export class SuiSearchValueAccessor<T extends RecursiveObject> implements ControlValueAccessor {
     onChange = () => {};
     onTouched = () => {};
 
-    constructor(private host:SuiSearch) {}
+    constructor(private host:SuiSearch<T>) {}
 
-    writeValue(value: any): void {
+    writeValue(value:T) {
         this.host.writeValue(value);
     }
 
-    registerOnChange(fn: () => void): void { this.onChange = fn; }
-    registerOnTouched(fn: () => void): void { this.onTouched = fn; }
+    registerOnChange(fn:() => void) {
+        this.onChange = fn;
+    }
+    registerOnTouched(fn:() => void) {
+        this.onTouched = fn;
+    }
 }
-
-export const SUI_SEARCH_DIRECTIVES = [SuiSearch, SuiSearchValueAccessor];
